remove AssemblerLib dependency by moving DoF related stuff in Math/LinAlg to AssemblerLib
