+++ AssemblerLib/GlobalMatrixProviders.cpp
